Rename SerDesDevice folder to serdes
