+++ BREAKING_CHANGES.md
@@ -1,3 +1,4 @@
 v1.1.4:
 
 * `USBDevice` folder renamed to `usb`
+* `SerDesDevice` folder renamed to `serdes`
