+++ tests/test_firmware_loopback/User/main.c
@@ -31,7 +31,7 @@ limitations under the License.
 #include "usb_device.h"
 #include "wch-ch56x-lib/HSPIDevice/hspi.h"
 #include "wch-ch56x-lib/logging/logging.h"
-#include "wch-ch56x-lib/SerDesDevice/serdes.h"
+#include "wch-ch56x-lib/serdes/serdes.h"
 #include "wch-ch56x-lib/usb/usb20.h"
 #include "wch-ch56x-lib/usb/usb30.h"
 #include "wch-ch56x-lib/usb/usb_descriptors.h"
